speed updating is working for android platform

+++ shared/src/commonMain/kotlin/com/apps/speed/domain/LiveSpeedProvider.kt
@@ -1,6 +1,6 @@
 package com.apps.speed.domain
 
 interface LiveSpeedProvider {
-    fun startSpeedUpdates(onSpeedChanged: (Int) -> Unit)
+    fun startSpeedUpdates(onSpeedChanged: (Float) -> Unit)
     fun stopSpeedUpdates()
 }
